pyproject.toml
[project]
name = "dataset-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Volker Lorrmann", email = "ligno.blades@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "fsspec>=2025.9.0",
    "numpy>=2.3.3",
    "polars>=1.34.0",
    "s3fs>=2025.9.0",
]

[project.optional-dependencies]
delta = ["deltalake>=1.1.4", "pyarrow>=16.1.0"]
iceberg = ["pyiceberg[sql]>=0.10.0", "pyarrow>=16.1.0"]
ducklake = ["duckdb>=1.4.1", "sqlalchemy>=2.0.0"]
cli = ["typer[all]>=0.12.5"]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[project.scripts]
dataset-generator = "dataset_generator.cli:main"

[tool.uv]
dev-dependencies = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[tool.hatch.envs.docs]
dependencies = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0", 
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build"
serve = "mkdocs serve"
deploy = "mkdocs gh-deploy"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "jupyterlab>=4.4.9",
    "marimo>=0.16.5",
    "datafusion>=50.0.0",
    "ibis-framework[duckdb]>=10.8.0",
    "ruff>=0.14.0",
    "openai>=2.3.0",
]
